change structure to remove puts from lib files, update the tests
