+++ bin/asr
@@ -1,5 +1,5 @@
 #!/usr/bin/env ruby
-require_relative '../lib/presenter'
+require_relative '../lib/array_builders'
 require 'colorize'
 
 # Here will be the main user interface and interaction.
@@ -36,17 +36,33 @@ puts
 
 # instantiate class
 
-display_deals = Presenter::OffersPresenter.new
+display_deals = ArrayBuilders::OffersBuilder.new
 
-# call the present_deals to show results
+# present results on screen, 5 at a time.
+@count = 0
 
-display_deals.present_deals(printing_mode: true)
+display_deals.offers_array_builder.each do |h|
+  puts "#{@count + 1}. " + h[:title].yellow.bold
+  puts h[:original_price].blue
+  puts h[:discount].green.bold
+  puts h[:price].green.bold
+  puts h[:url]
+  puts
+  @count += 1
+  next unless (@count % 5).zero?
+
+  puts
+  puts 'Press any key to continue...' unless @count == 40
+
+  $stdin.getch unless @count == 40
+  puts
+end
 
 puts
 puts
 puts "Prepare your wallet \u{1F4B8}  \u{1f607}"
-puts 'https://www.youtube.com/watch?v=rP2MDtWu5t0'.green.bold
 puts
+
 # the offers_array_builder method is called which reads the scraped data and puts it in '@deals', which is an array.
 
 # Say goodbye
