+++ lib/scrapper.rb
@@ -1,8 +1,7 @@
 require 'httparty'
 require 'nokogiri'
-require 'colorize'
-require 'io/console'
 require 'spinning_cursor'
+require 'io/console'
 
 module Scrapper
   class PageScrapper
@@ -20,30 +19,6 @@ module Scrapper
       parsed_offers = PageScrapper.page_scrapper.css('div#tab_specials_content a') # 40 deals
       SpinningCursor.stop
       parsed_offers
-    end
-
-    public
-
-    # put all the parsed offers in an array
-    def offers_array_builder
-      @deal_listings = offers_parser
-      @deals = []
-      i = 0
-      while i <= 39
-        @deal_listings.each do |deal_listing| # iterate through all listings and build a hash with data from each
-          @deal = {
-            title: deal_listing.css('div.tab_item_name').text,
-            original_price: deal_listing.css('div.discount_original_price').text,
-            discount: deal_listing.css('div.discount_pct').text,
-            price: deal_listing.css('div.discount_final_price').text,
-            url: @deal_listings.css('a')[i].attribute('href').value
-          }
-
-          @deals << @deal # put all deals into an array called deals.
-          i += 1
-        end
-      end
-      @deals[0..39] # return the first 40 array of deals
     end
   end
 end
