+++ spec/asr_spec.rb
@@ -1,5 +1,5 @@
 require_relative '../lib/scrapper'
-require_relative '../lib/presenter'
+require_relative '../lib/array_builders'
 require 'rspec'
 
 describe 'PageScraper' do
@@ -26,50 +26,28 @@ describe 'PageScraper' do
     end
   end
 
-  describe '#offers_array_builder' do
-    it 'Is not nil' do
-      expect(main_scrapper.offers_array_builder).not_to be_nil
-    end
-    it 'Returns objects' do
-      expect(main_scrapper.offers_array_builder).to be_a Object
-    end
-    it 'Returns an array' do
-      expect(main_scrapper.offers_array_builder).to be_an Array
-    end
-    it 'Has 40 elements' do
-      expect(main_scrapper.offers_array_builder.size).to eq(40)
-    end
-    it 'Has a key called "url" that includes "https"' do
-      expect(main_scrapper.offers_array_builder[0][:url]).to include('https')
-    end
-    it 'Has a key called "title"' do
-      expect(main_scrapper.offers_array_builder[0]).to have_key(:title)
-    end
-  end
-end
+  describe 'OffersBuilder' do
+    let(:build_results) { ArrayBuilders::OffersBuilder.new }
 
-describe 'OffersPresenter' do
-  let(:present_results) { Presenter::OffersPresenter.new }
-
-  # for testing purposes this one returns ['test']
-  describe 'present_deals' do
-    it 'Is an array' do
-      expect(present_results.present_deals).to eq ['test']
-    end
-    # Check true argument with mock data.
-    it 'takes true argument' do
-      dbl = double('mock', arg: true)
-      expect(dbl.arg).to eq(true)
-    end
-    # Check false argument with mock data.
-    it 'takes false argument' do
-      dbl2 = double('mock2', arg: false)
-      expect(dbl2.arg).to eq(false)
-    end
-    it 'test' do
-      dbl3 = double('takes #present_deals with a false argument for testing')
-      expect(dbl3).to receive(:present_deals).with(false)
-      dbl3.present_deals(false)
+    describe '#offers_array_builder' do
+      it 'Is not nil' do
+        expect(build_results.offers_array_builder).not_to be_nil
+      end
+      it 'Returns objects' do
+        expect(build_results.offers_array_builder).to be_a Object
+      end
+      it 'Returns an array' do
+        expect(build_results.offers_array_builder).to be_an Array
+      end
+      it 'Has 40 elements' do
+        expect(build_results.offers_array_builder.size).to eq(40)
+      end
+      it 'Has a key called "url" that includes "https"' do
+        expect(build_results.offers_array_builder[0][:url]).to include('https')
+      end
+      it 'Has a key called "title"' do
+        expect(build_results.offers_array_builder[0]).to have_key(:title)
+      end
     end
   end
 end
